+++ packages/bundler-plugin-core/test/sentry/logger.test.ts
@@ -1,44 +1,53 @@
 import { createLogger } from "../../src/sentry/logger";
 
 describe("Logger", () => {
-  const consoleLogSpy = jest.spyOn(console, "log").mockImplementation(() => undefined);
+  const consoleErrorSpy = jest.spyOn(console, "error").mockImplementation(() => undefined);
+  const consoleInfoSpy = jest.spyOn(console, "info").mockImplementation(() => undefined);
+  const consoleWarnSpy = jest.spyOn(console, "warn").mockImplementation(() => undefined);
+  const consoleDebugSpy = jest.spyOn(console, "debug").mockImplementation(() => undefined);
 
   afterEach(() => {
-    consoleLogSpy.mockReset();
+    consoleErrorSpy.mockReset();
+    consoleInfoSpy.mockReset();
+    consoleWarnSpy.mockReset();
+    consoleDebugSpy.mockReset();
   });
 
   it.each([
-    ["info", "Info"],
-    ["warn", "Warning"],
-    ["error", "Error"],
-  ] as const)(".%s() should log correctly", (loggerMethod, logLevel) => {
+    ["info", "Info", consoleInfoSpy],
+    ["warn", "Warning", consoleWarnSpy],
+    ["error", "Error", consoleErrorSpy],
+  ] as const)(".%s() should log correctly", (loggerMethod, logLevel, consoleSpy) => {
     const prefix = "[some-prefix]";
     const logger = createLogger({ prefix, silent: false, debug: true });
 
     logger[loggerMethod]("Hey!");
 
-    expect(consoleLogSpy).toHaveBeenCalledWith(`[some-prefix] ${logLevel}: Hey!`);
+    expect(consoleSpy).toHaveBeenCalledWith(`[some-prefix] ${logLevel}: Hey!`);
   });
 
   it.each([
-    ["info", "Info"],
-    ["warn", "Warning"],
-    ["error", "Error"],
-  ] as const)(".%s() should log multiple params correctly", (loggerMethod, logLevel) => {
-    const prefix = "[some-prefix]";
-    const logger = createLogger({ prefix, silent: false, debug: true });
+    ["info", "Info", consoleInfoSpy],
+    ["warn", "Warning", consoleWarnSpy],
+    ["error", "Error", consoleErrorSpy],
+  ] as const)(
+    ".%s() should log multiple params correctly",
+    (loggerMethod, logLevel, consoleSpy) => {
+      const prefix = "[some-prefix]";
+      const logger = createLogger({ prefix, silent: false, debug: true });
 
-    logger[loggerMethod]("Hey!", "this", "is", "a test with", 5, "params");
+      logger[loggerMethod]("Hey!", "this", "is", "a test with", 5, "params");
 
-    expect(consoleLogSpy).toHaveBeenCalledWith(
-      `[some-prefix] ${logLevel}: Hey!`,
-      "this",
-      "is",
-      "a test with",
-      5,
-      "params"
-    );
-  });
+      expect(consoleSpy).toHaveBeenCalledWith(
+        `[some-prefix] ${logLevel}: Hey!`,
+        "this",
+        "is",
+        "a test with",
+        5,
+        "params"
+      );
+    }
+  );
 
   it(".debug() should log correctly", () => {
     const prefix = "[some-prefix]";
@@ -46,7 +55,7 @@ describe("Logger", () => {
 
     logger.debug("Hey!");
 
-    expect(consoleLogSpy).toHaveBeenCalledWith(`[some-prefix] Debug: Hey!`);
+    expect(consoleDebugSpy).toHaveBeenCalledWith(`[some-prefix] Debug: Hey!`);
   });
 
   it(".debug() should log multiple params correctly", () => {
@@ -55,7 +64,7 @@ describe("Logger", () => {
 
     logger.debug("Hey!", "this", "is", "a test with", 5, "params");
 
-    expect(consoleLogSpy).toHaveBeenCalledWith(
+    expect(consoleDebugSpy).toHaveBeenCalledWith(
       `[some-prefix] Debug: Hey!`,
       "this",
       "is",
@@ -66,13 +75,17 @@ describe("Logger", () => {
   });
 
   describe("doesn't log when `silent` option is `true`", () => {
-    it.each(["info", "warn", "error"] as const)(".%s()", (loggerMethod) => {
+    it.each([
+      ["info", consoleInfoSpy],
+      ["warn", consoleWarnSpy],
+      ["error", consoleErrorSpy],
+    ] as const)(".%s()", (loggerMethod, consoleSpy) => {
       const prefix = "[some-prefix]";
       const logger = createLogger({ prefix, silent: true, debug: true });
 
       logger[loggerMethod]("Hey!");
 
-      expect(consoleLogSpy).not.toHaveBeenCalled();
+      expect(consoleSpy).not.toHaveBeenCalled();
     });
   });
 
@@ -82,6 +95,6 @@ describe("Logger", () => {
 
     logger.debug("Hey!");
 
-    expect(consoleLogSpy).not.toHaveBeenCalled();
+    expect(consoleDebugSpy).not.toHaveBeenCalled();
   });
 });